requirements.txt
langchain==0.1.0
streamlit==1.30.0
python-dotenv==1.0.0
pypdf==3.17.4
chromadb==0.4.22
streamlit_chat==0.1.1
faiss-cpu==1.7.4
pycryptodome==3.15.0
requests
azure-storage-blob
replicate
sentence-transformers